+++ opentelemetry-semantic-conventions/scripts/generate-consts-from-spec.sh
@@ -5,8 +5,8 @@ SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 CRATE_DIR="${SCRIPT_DIR}/../"
 
 # freeze the spec version and generator version to make generation reproducible
-SPEC_VERSION=1.26.0
-SEMCOVGEN_VERSION=0.24.0
+SPEC_VERSION=1.27.0
+SEMCOVGEN_VERSION=0.25.0
 
 cd "$CRATE_DIR"
 
@@ -35,7 +35,7 @@ docker run --rm \
 	-v "${CRATE_DIR}/scripts/templates:/templates" \
 	-v "${CRATE_DIR}/src:/output" \
 	otel/semconvgen:$SEMCOVGEN_VERSION \
-  --only span,event,scope \
+  --only span,event \
   -f /source code \
 	--template /templates/semantic_attributes.rs.j2 \
 	--output /output/trace.rs \